Enable output compilation validation in binder gen unit tests
